LightWallet can build_transaction
