+++ zingolib/src/lightclient/send.rs
@@ -109,76 +109,15 @@ pub mod send_with_proposal {
             &self,
             proposal: &Proposal<zcash_primitives::transaction::fees::zip317::FeeRule, NoteRef>,
         ) -> Result<NonEmpty<TxId>, CompleteAndBroadcastError> {
-            if self
-                .wallet
-                .transaction_context
-                .transaction_metadata_set
-                .read()
-                .await
-                .witness_trees()
-                .is_none()
-            {
-                return Err(CompleteAndBroadcastError::NoSpendCapability);
-            }
-
-            // Reset the progress to start. Any errors will get recorded here
-            self.wallet.reset_send_progress().await;
-
             let submission_height = self
                 .get_submission_height()
                 .await
                 .map_err(CompleteAndBroadcastError::SubmissionHeight)?;
 
-            let (sapling_output, sapling_spend): (Vec<u8>, Vec<u8>) =
-                read_sapling_params().map_err(CompleteAndBroadcastError::SaplingParams)?;
-            let sapling_prover = LocalTxProver::from_bytes(&sapling_spend, &sapling_output);
-            let unified_spend_key =
-                UnifiedSpendingKey::try_from(self.wallet.wallet_capability().as_ref())
-                    .map_err(CompleteAndBroadcastError::UnifiedSpendKey)?;
-
-            // We don't support zip320 yet. Only one step.
-            if proposal.steps().len() != 1 {
-                return Err(CompleteAndBroadcastError::ExchangeAddressesNotSupported);
-            }
-
-            let step = proposal.steps().first();
-
-            // The 'UnifiedSpendingKey' we create is not a 'proper' USK, in that the
-            // transparent key it contains is not the account spending key, but the
-            // externally-scoped derivative key. The goal is to fix this, but in the
-            // interim we use this special-case logic.
-            fn usk_to_tkey(
-                unified_spend_key: &UnifiedSpendingKey,
-                t_metadata: &TransparentAddressMetadata,
-            ) -> SecretKey {
-                hdwallet::ExtendedPrivKey::deserialize(&unified_spend_key.transparent().to_bytes())
-                    .expect("This a hack to do a type conversion, and will not fail")
-                    .derive_private_key(t_metadata.address_index().into())
-                    // This is unwrapped in librustzcash, so I'm not too worried about it
-                    .expect("private key derivation failed")
-                    .private_key
-            }
-
-            let build_result =
-                zcash_client_backend::data_api::wallet::calculate_proposed_transaction(
-                    self.wallet
-                        .transaction_context
-                        .transaction_metadata_set
-                        .write()
-                        .await
-                        .deref_mut(),
-                    &self.wallet.transaction_context.config.chain,
-                    &sapling_prover,
-                    &sapling_prover,
-                    &unified_spend_key,
-                    zcash_client_backend::wallet::OvkPolicy::Sender,
-                    proposal.fee_rule(),
-                    proposal.min_target_height(),
-                    &[],
-                    step,
-                    Some(usk_to_tkey),
-                    Some(self.wallet.wallet_capability().first_sapling_address()),
-                )?;
+            let build_result = self
+                .wallet
+                .build_transaction(proposal, submission_height)
+                .await?;
 
             let result = self
                 .wallet
